Split App::AutoTransaction into its own file
