+++ src/Mod/Spreadsheet/Gui/SheetTableView.cpp
@@ -31,6 +31,7 @@
 #endif
 
 #include <App/Application.h>
+#include <App/AutoTransaction.h>
 #include <App/Document.h>
 #include <Gui/Command.h>
 #include <Gui/MainWindow.h>
